+++ src/write_outputs/transmission/write_transmission_flows.jl
@@ -22,7 +22,7 @@ function write_transmission_flows(path::AbstractString,
         dfFlow = hcat(dfFlow, DataFrame(flow, :auto))
         auxNew_Names = [Symbol("Line"); [Symbol("t$t") for t in 1:T]]
         rename!(dfFlow, auxNew_Names)
-        CSV.write(filepath, dftranspose(dfFlow, false), writeheader = false)
+        write_transposed_csv(filepath, dfFlow, writeheader = false)
 
         if setup["OutputFullTimeSeries"] == 1 && setup["TimeDomainReduction"] == 1
             write_full_time_series_reconstruction(path, setup, dfFlow, "flow")